refactor(ffi): move sipi_init out of the oracle label into the seam
`sipi_init` — the production server-mode engine install the Rust shell calls
at startup — was defined in `src/cli/cli_app.cpp`, the component
`CONVENTIONS.md` labels oracle-only. That is a production FFI entry living
inside the oracle label (DUNE-003), while its sibling seam entries
(`sipi_imgroot`, `sipi_docroot`, …) already live in `src/ffi`.

BUILD-graph analysis: `src/cli` already depends on `src/ffi`, so the shared
startup helpers cannot stay behind in `src/cli` once `sipi_init` moves to
`src/ffi` (that would be a `cli↔ffi` cycle). They move with it; the dependency
direction stays cli → ffi (oracle → production).

Moved to `src/ffi` (homed by concern):
- `init.cpp`: `sipi_init` + its exclusive helpers (`parse_scaling_method`,
  `to_scaling_quality`, the `ServerRuntime` holder).
- `startup.{h,cpp}`: `LibraryInitialiser` (curl / Exiv2 XMP / TIFF global init)
  and `detect_available_memory` — shared with `sipi_cli_main` and `run_server`,
  now reached as `Sipi::ffi::…`.
- `sipiConfGlobals` (the `config` Lua-globals installer) folds into the existing
  `ffi/lua_config.{h,cpp}` next to the VM factory that consumes it.

`cli_app.cpp` shrinks by ~490 lines and calls the helpers through the seam.
No behaviour change: the differential parity gate passes, so the oracle and the
Rust shell still agree across the corpus. Also corrects the now-stale
"currently by SipiHttpServer; later by sipi_init" comment in
`engine_context.cpp` (both install the context now — production `sipi_init`,
oracle `SipiHttpServer`) at the single `set_engine_context` sink, and sharpens
`CONVENTIONS.md`'s oracle claim to `src/cli`'s **server mode** only.

Part of DEV-6968.
