+++ src/cli/cli_app.cpp
@@ -52,6 +52,7 @@
 #include "ffi/engine_context.h"
 #include "ffi/lua_config.h"
 #include "ffi/sipi_ffi.h"
+#include "ffi/startup.h"
 #include "SipiReport.h"
 #include "populate_from_image.h"
 #include "formats/SipiIOTiff.h"
@@ -121,59 +122,6 @@
   return (cores > 0) ? cores : 4;
 }
 
-/// Detect available memory from container limits or system info.
-/// Returns bytes, or 0 if detection fails.
-[[nodiscard]] static size_t detect_available_memory()
-{
-#ifdef __linux__
-  // 1. cgroups v2: /sys/fs/cgroup/memory.max
-  if (std::ifstream mem_max("/sys/fs/cgroup/memory.max"); mem_max.is_open()) {
-    std::string limit_str;
-    mem_max >> limit_str;
-    if (limit_str != "max") {
-      try {
-        return static_cast<size_t>(std::stoll(limit_str));
-      } catch (...) {
-        // Parse failure — fall through
-      }
-    }
-  }
-
-  // 2. cgroups v1: /sys/fs/cgroup/memory/memory.limit_in_bytes
-  if (std::ifstream mem_limit("/sys/fs/cgroup/memory/memory.limit_in_bytes"); mem_limit.is_open()) {
-    long long limit = 0;
-    mem_limit >> limit;
-    // 9223372036854771712 = kernel "unlimited" sentinel
-    if (limit > 0 && limit < 9223372036854771712LL) {
-      return static_cast<size_t>(limit);
-    }
-  }
-
-  // 3. /proc/meminfo fallback
-  if (std::ifstream meminfo("/proc/meminfo"); meminfo.is_open()) {
-    std::string line;
-    while (std::getline(meminfo, line)) {
-      if (line.rfind("MemTotal:", 0) == 0) {
-        // Format: "MemTotal:     16384000 kB"
-        long long kb = 0;
-        std::sscanf(line.c_str(), "MemTotal: %lld kB", &kb);
-        if (kb > 0) return static_cast<size_t>(kb) * 1024;
-      }
-    }
-  }
-#endif
-
-#ifdef __APPLE__
-  // macOS: sysctl hw.memsize
-  int64_t memsize = 0;
-  size_t len = sizeof(memsize);
-  if (sysctlbyname("hw.memsize", &memsize, &len, nullptr, 0) == 0 && memsize > 0) {
-    return static_cast<size_t>(memsize);
-  }
-#endif
-
-  return 0;
-}
 
 /*!
  * \mainpage
@@ -204,138 +152,6 @@
  *
  */
 
-static void sipiConfGlobals(lua_State *L, shttps::RequestContext &ctx, void *user_data)
-{
-  auto *conf = (Sipi::SipiConf *)user_data;
-
-  lua_createtable(L, 0, 14);// table1
-
-  lua_pushstring(L, "hostname");// table1 - "index_L1"
-  lua_pushstring(L, conf->getHostname().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "port");// table1 - "index_L1"
-  lua_pushinteger(L, conf->getPort());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "sslport");// table1 - "index_L1"
-  lua_pushinteger(L, conf->getSSLPort());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "imgroot");// table1 - "index_L1"
-  lua_pushstring(L, conf->getImgRoot().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "max_temp_file_age");// table1 - "index_L1"
-  lua_pushinteger(L, conf->getMaxTempFileAge());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "prefix_as_path");// table1 - "index_L1"
-  lua_pushboolean(L, conf->getPrefixAsPath());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "init_script");// table1 - "index_L1"
-  lua_pushstring(L, conf->getInitScript().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "cache_dir");// table1 - "index_L1"
-  lua_pushstring(L, conf->getCacheDir().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "cache_size");// table1 - "index_L1"
-  lua_pushinteger(L, conf->getCacheSize());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "jpeg_quality");// table1 - "index_L1"
-  lua_pushinteger(L, conf->getJpegQuality());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "keep_alive");// table1 - "index_L1"
-  lua_pushinteger(L, conf->getKeepAlive());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "thumb_size");// table1 - "index_L1"
-  lua_pushstring(L, conf->getThumbSize().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "cache_n_files");// table1 - "index_L1"
-  lua_pushinteger(L, conf->getCacheNFiles());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "n_threads");// table1 - "index_L1"
-  lua_pushinteger(L, conf->getNThreads());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "max_post_size");// table1 - "index_L1"
-  lua_pushinteger(L, conf->getMaxPostSize());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "tmpdir");// table1 - "index_L1"
-  lua_pushstring(L, conf->getTmpDir().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "ssl_certificate");// table1 - "index_L1"
-  lua_pushstring(L, conf->getSSLCertificate().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "ssl_key");// table1 - "index_L1"
-  lua_pushstring(L, conf->getSSLKey().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "scriptdir");// table1 - "index_L1"
-  lua_pushstring(L, conf->getScriptDir().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "logfile");// table1 - "index_L1"
-  lua_pushstring(L, conf->getLogfile().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "loglevel");// table1 - "index_L1"
-  const std::string loglevel = conf->getLoglevel();
-  if (loglevel == "EMERG") {
-    lua_pushinteger(L, LL_EMERG);
-  } else if (loglevel == "ALERT") {
-    lua_pushinteger(L, LL_ALERT);
-  } else if (loglevel == "CRIT") {
-    lua_pushinteger(L, LL_CRIT);
-  } else if (loglevel == "ERR") {
-    lua_pushinteger(L, LL_ERR);
-  } else if (loglevel == "WARNING") {
-    lua_pushinteger(L, LL_WARNING);
-  } else if (loglevel == "NOTICE") {
-    lua_pushinteger(L, LL_NOTICE);
-  } else if (loglevel == "INFO") {
-    lua_pushinteger(L, LL_INFO);
-  } else if (loglevel == "DEBUG") {
-    lua_pushinteger(L, LL_DEBUG);
-  } else {
-    lua_pushinteger(L, -1);
-  }
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "knora_path");// table1 - "index_L1"
-  lua_pushstring(L, conf->getKnoraPath().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "knora_port");// table1 - "index_L1"
-  lua_pushstring(L, conf->getKnoraPort().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "adminuser");// table1 - "index_L1"
-  lua_pushstring(L, conf->getAdminUser().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_pushstring(L, "password");// table1 - "index_L1"
-  lua_pushstring(L, conf->getPassword().c_str());
-  lua_rawset(L, -3);// table1
-
-  // TODO: in the sipi config file, there are different namespaces that are unified here (danger of collision)
-  lua_pushstring(L, "docroot");// table1 - "index_L1"
-  lua_pushstring(L, conf->getDocRoot().c_str());
-  lua_rawset(L, -3);// table1
-
-  lua_setglobal(L, "config");
-}
 
 // small function to check if file exist
 inline bool exists_file(const std::string &name)
@@ -344,313 +160,6 @@ inline bool exists_file(const std::string &name)
   return (stat(name.c_str(), &buffer) == 0);
 }
 
-/*!
- * A singleton that does global initialisation and cleanup of libraries used by Sipi. This class is used only
- * in main().
- *
- * Some libraries such as the ones used for processing XMP or TIFF need some global initialization. These
- * are performed in the constructor of this singleton class
- */
-class LibraryInitialiser
-{
-public:
-  /*!
-   * @return the singleton instance.
-   */
-  static LibraryInitialiser &instance()
-  {
-    // In C++11, initialization of this static local variable happens once and is thread-safe.
-    static LibraryInitialiser sipi_init;
-    return sipi_init;
-  }
-
-private:
-  LibraryInitialiser()
-  {
-    // Initialise libcurl.
-    curl_global_init(CURL_GLOBAL_ALL);
-
-    // Initialise Exiv2, registering namespace sipi. Since this is not thread-safe, it must
-    // be done here in the main thread.
-    if (!Exiv2::XmpParser::initialize(Sipi::xmplock_func, &Sipi::xmp_mutex)) {
-      throw shttps::Error("Exiv2::XmpParser::initialize failed");
-    }
-
-    // Inititalise the TIFF library.
-    Sipi::SipiIOTiff::initLibrary();
-  }
-
-  ~LibraryInitialiser()
-  {
-    // Clean up libcurl.
-    curl_global_cleanup();
-
-    // Clean up Exiv2.
-    Exiv2::XmpParser::terminate();
-  }
-};
-
-namespace {
-/*! Map a config scaling-quality string to a ScalingMethod; unknown/missing → HIGH
- *  (matching the legacy SipiHttpServer::scaling_quality setter). */
-Sipi::ScalingMethod parse_scaling_method(const std::string &v)
-{
-  if (v == "medium") { return Sipi::ScalingMethod::MEDIUM; }
-  if (v == "low") { return Sipi::ScalingMethod::LOW; }
-  return Sipi::ScalingMethod::HIGH;
-}
-
-/*! Convert SipiConf's `map<string,string>` scaling-quality table into the
- *  `ScalingQuality` struct EngineContext holds — the conversion the legacy
- *  SipiHttpServer::scaling_quality(map) setter performs (jk2 ← the "jpk" key). */
-Sipi::ScalingQuality to_scaling_quality(const std::map<std::string, std::string> &m)
-{
-  const auto get = [&](const char *k) -> std::string {
-    const auto it = m.find(k);
-    return it != m.end() ? it->second : std::string();
-  };
-  Sipi::ScalingQuality q;
-  q.jk2 = parse_scaling_method(get("jpk"));
-  q.jpeg = parse_scaling_method(get("jpeg"));
-  q.tiff = parse_scaling_method(get("tiff"));
-  q.png = parse_scaling_method(get("png"));
-  return q;
-}
-
-/*!
- * Process-wide server runtime installed by `sipi_init` (strangler-fig rewrite).
- *
- * The Rust shell, unlike the C++ `SipiHttpServer`, has no server object to own
- * the engine services, so `sipi_init` parks them here for the process lifetime.
- * `engine_context()` stores non-owning pointers into this holder, so it must
- * outlive every serve call — hence file-static. The held `SipiConf` also backs
- * the `sipiConfGlobals` installer captured in the Lua config (the per-request VM
- * factory reads it on every preflight / route call).
- */
-struct ServerRuntime
-{
-  Sipi::SipiConf conf;
-  std::unique_ptr<Sipi::SipiCache> cache;
-  std::unique_ptr<Sipi::SipiMemoryBudget> memory_budget;
-};
-std::unique_ptr<ServerRuntime> g_server_runtime;
-}// namespace
-
-/*!
- * Parse the Lua config and install the engine + Lua config from scratch
- * (strangler-fig rewrite). This is the from-scratch counterpart to the C++
- * server's parity install (`SipiHttpServer`'s `set_engine_context` +
- * run_server's `set_lua_config`); the Rust shell calls it once at startup
- * before serving. Builds the cache / memory budget into
- * `g_server_runtime`, points `engine_context()` at them, and installs the
- * engine-held Lua config VM factory. Returns 0 on success or `EXIT_FAILURE`;
- * never lets a C++ exception cross the boundary.
- *
- * `overrides` carries the CLI/env flags the Rust shell parsed (or null = none).
- * Present overrides are layered onto the Lua-parsed SipiConf below, before the
- * engine services read it. Only engine-behaviour flags are forwarded; transport
- * flags the Rust shell owns (TLS, keep-alive, concurrency) are not in the struct.
- */
-extern "C" int sipi_init(const char *lua_config_path, const SipiServerConfig *overrides)
-{
-  try {
-    // A null/empty config path selects the Lua-less init path: a TOML config,
-    // parsed Rust-side, supplies every value through `overrides`. Otherwise the
-    // Lua config is the base the overrides layer onto.
-    const bool has_lua_config = (lua_config_path != nullptr && lua_config_path[0] != '\0');
-
-    // Initialise the codec libraries (curl / Exiv2 / TIFF) the decode pipeline
-    // needs. The Rust server path does not go through sipi_cli_main's
-    // LibraryInitialiser, so sipi_init owns it here; the singleton is idempotent.
-    LibraryInitialiser::instance();
-
-    auto runtime = std::make_unique<ServerRuntime>();
-
-    // Parse the Lua config — the same VM the C++ server path reads via SipiConf.
-    // Skipped on the Lua-less path: `runtime->conf` stays default-constructed (an
-    // all-defaults config via SipiConf's in-class initializers) and the overrides
-    // below supply imgroot / scriptdir / etc.
-    if (has_lua_config) {
-      shttps::LuaServer luacfg(lua_config_path);
-      runtime->conf = Sipi::SipiConf(luacfg);
-    }
-    Sipi::SipiConf &conf = runtime->conf;
-
-    // CLI/env overrides: layer the present overrides onto the
-    // Lua-parsed SipiConf BEFORE the cache / memory-budget
-    // services below are built from `conf`, so an override reaches the engine.
-    // Setter names are SipiConf's verbatim (incl. the `setPasswort` typo). Sized
-    // strings (cache_size/maxpost/max_decode_memory) carry the raw "300M" text;
-    // parseSizeString expands the suffix engine-side. A negative maxpost /
-    // max-decode-memory clamps to 0 (matching the SipiConf ctor + run_server) so
-    // it cannot become SIZE_MAX and skip the budget auto-detect; cache_size keeps
-    // -1 as its valid "unlimited" sentinel. cache_nfiles is `unsigned` (0 =
-    // unlimited; a negative is rejected by CLI11 on both binaries / by clap u32),
-    // so it forwards straight to setCacheNFiles(size_t) with no signed wrap.
-    if (overrides != nullptr) {
-      const SipiServerConfig &o = *overrides;
-      // Strings (null = absent).
-      if (o.imgroot != nullptr) conf.setImgRoot(o.imgroot);
-      if (o.scriptdir != nullptr) conf.setScriptDir(o.scriptdir);
-      if (o.initscript != nullptr) conf.setInitScript(o.initscript);
-      if (o.tmpdir != nullptr) conf.setTmpDir(o.tmpdir);
-      if (o.jwtkey != nullptr) conf.setJwtSecret(o.jwtkey);
-      if (o.adminuser != nullptr) conf.setAdminUser(o.adminuser);
-      if (o.adminpasswd != nullptr) conf.setPasswort(o.adminpasswd);// `setPasswort` is the real (typo'd) setter
-      if (o.cache_dir != nullptr) conf.setCacheDir(o.cache_dir);
-      if (o.cache_size != nullptr) conf.setCacheSize(Sipi::parseSizeString(o.cache_size));
-      if (o.maxpost != nullptr) {
-        const long long v = Sipi::parseSizeString(o.maxpost);
-        conf.setMaxPostSize(v > 0 ? static_cast<size_t>(v) : 0);// <=0 -> 0 (unlimited)
-      }
-      if (o.max_decode_memory != nullptr) {
-        const long long v = Sipi::parseSizeString(o.max_decode_memory);
-        conf.setMaxDecodeMemory(v > 0 ? static_cast<size_t>(v) : 0);// <=0 -> 0 (auto-detect 75% RAM)
-      }
-      if (o.decode_memory_mode != nullptr) conf.setDecodeMemoryMode(o.decode_memory_mode);
-      if (o.thumbsize != nullptr) conf.setThumbSize(o.thumbsize);
-      if (o.knorapath != nullptr) conf.setKnoraPath(o.knorapath);
-      if (o.knoraport != nullptr) conf.setKnoraPort(o.knoraport);
-      if (o.docroot != nullptr) conf.setDocRoot(o.docroot);
-      if (o.wwwroute != nullptr) conf.setWWWRoute(o.wwwroute);
-      if (o.loglevel != nullptr) conf.setLogLevel(o.loglevel);
-      // Scaling-quality per codec (TOML-config-only — no CLI flag). Merge the
-      // present codecs onto the base map so a partial override keeps the others.
-      // The "j2k" key is stored as the config writes it; to_scaling_quality reads
-      // it under "jpk" (a legacy engine quirk), so j2k scaling falls to the
-      // default on both the Lua and TOML paths alike — parity, not a new bug.
-      if (o.scaling_quality_jpeg != nullptr || o.scaling_quality_tiff != nullptr
-          || o.scaling_quality_png != nullptr || o.scaling_quality_j2k != nullptr) {
-        std::map<std::string, std::string> sq = conf.getScalingQuality();
-        if (o.scaling_quality_jpeg != nullptr) sq["jpeg"] = o.scaling_quality_jpeg;
-        if (o.scaling_quality_tiff != nullptr) sq["tiff"] = o.scaling_quality_tiff;
-        if (o.scaling_quality_png != nullptr) sq["png"] = o.scaling_quality_png;
-        if (o.scaling_quality_j2k != nullptr) sq["j2k"] = o.scaling_quality_j2k;
-        conf.setScalingQuality(sq);
-      }
-      if (o.subdirexcludes != nullptr && o.subdirexcludes_len > 0) {
-        std::vector<std::string> excludes;
-        excludes.reserve(o.subdirexcludes_len);
-        for (size_t i = 0; i < o.subdirexcludes_len; ++i) excludes.emplace_back(o.subdirexcludes[i]);
-        conf.setSubdirExcludes(excludes);
-      }
-      // Scalars (presence flag — 0 is a valid value, so gate on has_).
-      if (o.has_serverport) conf.setPort(o.serverport);
-      if (o.has_maxtmpage) conf.setMaxTempFileAge(o.maxtmpage);
-      if (o.has_cache_nfiles) conf.setCacheNFiles(o.cache_nfiles);
-      if (o.has_subdirlevels) conf.setSubdirLevels(o.subdirlevels);
-      if (o.has_pathprefix) conf.setPrefixAsPath(o.pathprefix != 0);
-      if (o.has_max_pixel_limit) conf.setMaxPixelLimit(o.max_pixel_limit);
-      if (o.has_jpeg_quality) conf.setJpegQuality(o.jpeg_quality);
-    }
-
-    // Engine services, mirroring run_server()'s construction (config values, with
-    // the CLI/env overrides above already applied). A null service means
-    // the corresponding feature is disabled, matching the legacy accessors.
-    {
-      const std::string cachedir = conf.getCacheDir();
-      const long long cache_size = conf.getCacheSize();
-      if (cache_size != 0 && !cachedir.empty()) {
-        // Degrade to no-cache on a bad/unwritable cache dir rather than aborting
-        // startup — parity with SipiHttpServer::cache() (which logs + continues).
-        try {
-          runtime->cache = std::make_unique<Sipi::SipiCache>(cachedir, cache_size, conf.getCacheNFiles());
-        } catch (const shttps::Error &e) {
-          log_warn("sipi_init: caching disabled — %s", e.what());
-          runtime->cache = nullptr;
-        }
-      }
-    }
-    {
-      const Sipi::MemoryBudgetMode mode = Sipi::parse_memory_budget_mode(conf.getDecodeMemoryMode());
-      if (mode != Sipi::MemoryBudgetMode::OFF) {
-        std::size_t budget = conf.getMaxDecodeMemory();
-        if (budget == 0) {
-          const std::size_t detected = detect_available_memory();
-          budget = (detected > 0) ? detected * 3 / 4 : (1ULL * 1024 * 1024 * 1024);
-        }
-        runtime->memory_budget = std::make_unique<Sipi::SipiMemoryBudget>(budget, mode);
-        Sipi::observability::Metrics::instance().decode_memory_budget_bytes.Set(static_cast<double>(budget));
-      }
-    }
-
-    // Resolve the image root (realpath) for path-traversal containment (R2),
-    // mirroring SipiHttpServer's resolve at startup.
-    const std::string imgroot = conf.getImgRoot();
-    char resolved[PATH_MAX];
-    if (realpath(imgroot.c_str(), resolved) == nullptr) {
-      log_err("sipi_init: image root '%s' does not resolve", imgroot.c_str());
-      return EXIT_FAILURE;
-    }
-    const std::string resolved_imgroot(resolved);
-
-    // Read the init script — the last fallible step — BEFORE any install. The
-    // engine context and Lua config install non-owning pointers into `runtime`;
-    // installing them first and then failing here would leave the file-static
-    // engine pointing into `runtime`, which is freed on this early return.
-    std::string initscript_src;
-    if (!conf.getInitScript().empty()) {
-      std::ifstream initscript_in(conf.getInitScript());
-      if (initscript_in.fail()) {
-        log_err("sipi_init: initscript \"%s\" not found", conf.getInitScript().c_str());
-        return EXIT_FAILURE;
-      }
-      initscript_src.assign(
-        (std::istreambuf_iterator<char>(initscript_in)), std::istreambuf_iterator<char>());
-    }
-    // else (Lua-less / TOML, no initscript set): leave initscript_src empty —
-    // set_lua_config accepts it and the sipiGlobals installers still register the
-    // `server` Lua table, so configured route scripts run.
-
-    // Install the engine context — non-owning pointers into g_server_runtime.
-    Sipi::ffi::set_engine_context(Sipi::ffi::EngineContext{
-      .cache = runtime->cache.get(),
-      .memory_budget = runtime->memory_budget.get(),
-      .imgroot = imgroot,
-      .resolved_imgroot = resolved_imgroot,
-      .docroot = conf.getDocRoot(),
-      .wwwroute = conf.getWWWRoute(),
-      .prefix_as_path = conf.getPrefixAsPath(),
-      .jpeg_quality = conf.getJpegQuality(),
-      .scaling_quality = to_scaling_quality(conf.getScalingQuality()),
-      .max_pixel_limit = conf.getMaxPixelLimit(),
-      .nthreads = static_cast<int>(conf.getNThreads()),
-      .port = conf.getPort(),
-      .max_post_size = conf.getMaxPostSize(),
-    });
-
-    // Install the engine-held Lua config (the per-call VM factory behind
-    // sipi_preflight / sipi_run_lua_route). Same scriptdir, JWT secret, and
-    // globals installers (in registration order) as run_server's set_lua_config.
-    // The sipiConfGlobals installer captures &conf, which stays valid: `runtime`
-    // is heap-allocated, so its SipiConf address is stable across the move into
-    // g_server_runtime below.
-    Sipi::ffi::set_lua_config(Sipi::ffi::LuaConfig{
-      .init_script = std::move(initscript_src),
-      .script_dir = conf.getScriptDir(),
-      .jwt_secret = conf.getJwtSecret(),
-      .globals = {
-        { sipiConfGlobals, &conf },
-        { shttps::sqliteGlobals, nullptr },
-        { Sipi::sipiGlobals, nullptr },
-      },
-      .routes = conf.getRoutes(),
-    });
-
-    g_server_runtime = std::move(runtime);
-    log_info("sipi_init: engine + Lua config installed (imgroot resolved: %s)", resolved_imgroot.c_str());
-    return EXIT_SUCCESS;
-  } catch (const shttps::Error &e) {
-    log_err("sipi_init failed: %s", e.what());
-    return EXIT_FAILURE;
-  } catch (const std::exception &e) {
-    log_err("sipi_init failed: %s", e.what());
-    return EXIT_FAILURE;
-  } catch (...) {
-    log_err("sipi_init failed: unknown error");
-    return EXIT_FAILURE;
-  }
-}
 
 /*!
  * The CLI entry point, behind the FFI seam (`ffi/sipi_ffi.h`).
@@ -681,8 +190,8 @@ extern "C" int sipi_cli_main(int argc, char **argv)
   // first we initialize the libraries that sipi uses
   //
   try {
-    LibraryInitialiser &sipi_init = LibraryInitialiser::instance();
-    _unused(sipi_init);// Silence compiler warning about unused variable.
+    Sipi::ffi::LibraryInitialiser &lib_init = Sipi::ffi::LibraryInitialiser::instance();
+    _unused(lib_init);// Silence compiler warning about unused variable.
   } catch (shttps::Error &e) {
     log_err("Library initialization failed: %s", e.to_string().c_str());
     return EXIT_FAILURE;
@@ -1387,7 +896,7 @@ extern "C" int sipi_cli_main(int argc, char **argv)
       server.max_post_size(sipiConf.getMaxPostSize());
       server.scriptdir(sipiConf.getScriptDir());
       server.luaRoutes(sipiConf.getRoutes());
-      server.add_lua_globals_func(sipiConfGlobals, &sipiConf);
+      server.add_lua_globals_func(Sipi::ffi::sipiConfGlobals, &sipiConf);
       server.add_lua_globals_func(shttps::sqliteGlobals);
       server.add_lua_globals_func(Sipi::sipiGlobals);
       server.prefix_as_path(sipiConf.getPrefixAsPath());
@@ -1416,7 +925,7 @@ extern "C" int sipi_cli_main(int argc, char **argv)
         auto mb_mode = Sipi::parse_memory_budget_mode(mb_mode_str);
         if (mb_mode != Sipi::MemoryBudgetMode::OFF) {
           if (budget_bytes == 0) {
-            size_t detected = detect_available_memory();
+            size_t detected = Sipi::ffi::detect_available_memory();
             if (detected > 0) {
               budget_bytes = detected * 3 / 4;
               log_info("Auto-detected memory: %zu bytes, budget set to 75%%: %zu bytes", detected, budget_bytes);
@@ -1481,7 +990,7 @@ extern "C" int sipi_cli_main(int argc, char **argv)
           .script_dir = sipiConf.getScriptDir(),
           .jwt_secret = sipiConf.getJwtSecret(),
           .globals = {
-            { sipiConfGlobals, &sipiConf },
+            { Sipi::ffi::sipiConfGlobals, &sipiConf },
             { shttps::sqliteGlobals, nullptr },
             { Sipi::sipiGlobals, nullptr },
           },
