+++ src/ffi/engine_context.cpp
@@ -10,13 +10,19 @@
 
 namespace Sipi::ffi {
 namespace {
-  // Process-wide engine state. Set once at startup (currently by SipiHttpServer;
-  // later by sipi_init) and only read thereafter, so no synchronization is
-  // needed for the read path the serve functions take.
+  // Process-wide engine state. Installed once at startup and only read
+  // thereafter, so no synchronization is needed for the read path the serve
+  // functions take.
   EngineContext g_engine;
   bool g_engine_installed = false;
 }// namespace
 
+// Two callers install the context, one per surface: in production, `sipi_init`
+// (`ffi/init.cpp`); in the retained oracle transport, `SipiHttpServer` at
+// startup (its parity install). Both run once, before any serve call, and write
+// the same shape — the oracle path exists only for the differential gate and
+// dies with the C++ server. Documented here at the single sink rather than at
+// each caller.
 void set_engine_context(const EngineContext &ctx)
 {
   g_engine = ctx;
